split main.cpp into client.cpp and server.cpp
