+++ misc.cpp
@@ -16,6 +16,9 @@
 int hb_mode=1;
 int hb_len=1200;
 
+char hb_buf[buf_len];
+
+
 int mtu_warn=1375;//if a packet larger than mtu warn is receviced,there will be a warning
 
 int max_rst_to_show=15;
